+++ src/app/app.component.html
@@ -1,4 +1,4 @@
-<div class="container">
-    <h1>Angular Junior </h1><h1>Сегодняшняя дата: {{ formattedDate}}</h1>
+<div class="header__container container">
+    <h1>Angular Junior </h1><h1>Дата сегодня: {{ formattedDate}}</h1>
     <router-outlet></router-outlet>
 </div>
