+++ src/app/app.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { UsersListComponent } from './users-list/users-list.component';
+import { UsersListComponent } from './components/users-list/users-list.component';
 import { TodosListComponent } from './todo-list/todos-list.component';
 
 export const routes: Routes = [
