+++ src/app/components/todo-list/todo-card/todo-card.component.scss
