+++ src/app/components/users-list/user-card/user-card.component.scss
