+++ src/app/components/users-list/users-list.component.ts
@@ -1,15 +1,14 @@
 import { AsyncPipe, NgFor } from "@angular/common";
 import { ChangeDetectionStrategy, Component, inject, input } from "@angular/core";
-import { UsersApiService } from "../service/users-api.service";
-import { UserCardComponent } from "./user-card/user-card.component";
-import { UserService } from "../service/users.service";
-import { CreateUserFormComponent } from "../create-user-form/create-user-form.component";
-import { IUser, ICreateUser } from "../interface/user.interface";
+import { UsersApiService } from "../../service/users-api.service";
+import { UserCardComponent } from "../users-list/user-card/user-card.component";
+import { UserService } from "../../service/users.service";
+import { CreateUserFormComponent } from "./create-user-form/create-user-form.component";
+import { IUser, ICreateUser } from "../../interface/user.interface";
 import { MatButton, MatButtonModule } from "@angular/material/button";
 import { MatIconModule } from "@angular/material/icon";
 import { MatDialog } from "@angular/material/dialog";
-import { CreateUserDialogComponent } from "./create-user-dialog/create-user-dialog.component";
-
+import { CreateUserDialogComponent } from "../users-list/create-user-dialog/create-user-dialog.component";
 
 @Component({
     selector: 'app-user-list',
